pyproject.toml
[project]
name = "wldd-pricing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "black>=25.1.0",
    "dotenv>=0.9.9",
    "numpy>=2.3.2",
    "openai>=1.106.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
]
